Moves reference unittests to separate file
